Support JSON Google-Secret Manager Encoding
Google Secret Manager's API returns secrets always as a `[]byte`.
Kubernetes secrets can have additional depth with multiple key and value
pairs.  To support this naturally within Pentagon, introduce a
GSMEncodingType parameter that allows the user to parse the contents of
the secret as a JSON file into a map[string]json.RawMessage, preserving
the encoding of the inner secret's value.  Then map each of those keys
to individual "sub-secrets" in K8s.