Support Unwrapping with gsmEncoding: json
When using `gsmEncoding: json` we want to unwrap the first level of
value so that extra quotes don't get included in the secret's value
during reflection.

So we want:
```
{"key1": "val1"}
```
...to become a secret with
key = "key1"
value = val1 (without the surrounding quotes)

Same for numbers.

Anything not a string or number will be remarshaled to json and saved as
the secret's value.